Refactor helped functions into separate file
This should fix the dependencies array and cause it to update when a planner name is changed.
